+++ snake/source/common/objects/Entity.h
@@ -10,7 +10,7 @@ namespace Engine
 
 class Entity {
 public:
-	Entity(Utilities::Point<float> position);
+	Entity(Utilities::Point<int> position);
 	void SetTexture(Engine::Texture2D texture) ;
 	void Update(int screenWidth, int screenHeight);
 	std::shared_ptr<Engine::Sprite> GetSprite() const { return mSprite; }
