+++ snake/source/common/objects/PointCollider.cpp
@@ -2,7 +2,7 @@
 
 using namespace Utilities;
 
-bool PointCollider::Collides(Point<float> point1, Point<float> point2)
+bool PointCollider::Collides(Point<int> point1, Point<int> point2)
 {
     return ( point1.X == point2.X && point1.Y == point2.Y);
 }
