+++ snake/source/common/objects/PointCollider.h
@@ -3,5 +3,5 @@
 
 class PointCollider {
 public:
-	static bool Collides(Utilities::Point<float> point1, Utilities::Point<float> point2);
+	static bool Collides(Utilities::Point<int> point1, Utilities::Point<int> point2);
 };
